Prevent Exception when trying to delete snapshots when there are none
If one calls the `snapshot:delete` command without providing a snapshot name and if there is no available snapshot, a \LogicException will be thrown when trying to build a list of choices.

This guards against the exception in a similar way to what is done within the `snapshot:list` and `snapshot:load` commands.